refactor: config에서 시간 변환 파일 분리, 자잘한 코드 리팩토링
