+++ src/main/java/com/smooth/accident_service/global/config/DynamoDBConfig.java
@@ -7,25 +7,19 @@ import org.springframework.context.annotation.Configuration;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
-import software.amazon.awssdk.enhanced.dynamodb.AttributeConverter;
-import software.amazon.awssdk.enhanced.dynamodb.AttributeValueType;
-import software.amazon.awssdk.enhanced.dynamodb.EnhancedType;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
-import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
-
-import java.time.LocalDateTime;
 
 @Configuration
 public class DynamoDBConfig {
 
-    @Value("${region:}")
+    @Value("${cloud.aws.region.static}")
     private String region;
 
-    @Value("${access-key}")
+    @Value("${cloud.aws.credentials.access-key}")
     private String accessKey;
 
-    @Value("${secret-key}")
+    @Value("${cloud.aws.credentials.secret-key}")
     private String secretKey;
 
     @Bean
@@ -43,27 +37,5 @@ public class DynamoDBConfig {
         return DynamoDbEnhancedClient.builder()
                 .dynamoDbClient(dynamoDbClient)
                 .build();
-    }
-
-    public static class LocalDateTimeConverter implements AttributeConverter<LocalDateTime> {
-        @Override
-        public AttributeValue transformFrom(LocalDateTime input) {
-            return AttributeValue.builder().s(input.toString()).build();
-        }
-
-        @Override
-        public LocalDateTime transformTo(AttributeValue input) {
-            return LocalDateTime.parse(input.s());
-        }
-
-        @Override
-        public EnhancedType<LocalDateTime> type() {
-            return EnhancedType.of(LocalDateTime.class);
-        }
-
-        @Override
-        public AttributeValueType attributeValueType() {
-            return AttributeValueType.S;
-        }
     }
 }
