+++ src/main/java/com/smooth/accident_service/global/dynamoDb/LocalDateTimeConverter.java
@@ -1,0 +1,33 @@
+package com.smooth.accident_service.global.dynamoDb;
+
+import software.amazon.awssdk.enhanced.dynamodb.AttributeConverter;
+import software.amazon.awssdk.enhanced.dynamodb.AttributeValueType;
+import software.amazon.awssdk.enhanced.dynamodb.EnhancedType;
+import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
+
+import java.time.LocalDateTime;
+
+public class LocalDateTimeConverter implements AttributeConverter<LocalDateTime> {
+
+    @Override
+    public AttributeValue transformFrom(LocalDateTime input) {
+        return AttributeValue.builder().s(input.toString()).build();
+    }
+
+    @Override
+    public LocalDateTime transformTo(AttributeValue input) {
+        return LocalDateTime.parse(input.s());
+    }
+
+    @Override
+    public EnhancedType<LocalDateTime> type() {
+        return EnhancedType.of(LocalDateTime.class);
+    }
+
+    @Override
+    public AttributeValueType attributeValueType() {
+        return AttributeValueType.S;
+    }
+}
+
+
